+++ src/main/java/com/microservices/choreographyorchestrator/domain/SagaStepTransitionEvent.java
@@ -1,5 +1,6 @@
 package com.microservices.choreographyorchestrator.domain;
 
+import com.microservices.choreographyorchestrator.domain.definition.SagaStepDefinition;
 import lombok.Getter;
 import lombok.Setter;
 import org.neo4j.ogm.annotation.*;
