+++ src/main/java/com/microservices/choreographyorchestrator/domain/definition/ExecutionInfo.java
@@ -1,4 +1,4 @@
-package com.microservices.choreographyorchestrator.domain.meta;
+package com.microservices.choreographyorchestrator.domain.definition;
 
 import lombok.Getter;
 import lombok.Setter;
