+++ src/main/java/com/microservices/choreographyorchestrator/dto/definition/FailExecutionInfoDto.java
@@ -1,11 +1,11 @@
-package com.microservices.choreographyorchestrator.dto.meta;
+package com.microservices.choreographyorchestrator.dto.definition;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class SuccessExecutionInfoDto {
+public class FailExecutionInfoDto {
     private String channelName;
 
     private String eventType;
