+++ src/main/java/com/microservices/choreographyorchestrator/dto/definition/SuccessExecutionInfoDto.java
@@ -1,11 +1,11 @@
-package com.microservices.choreographyorchestrator.dto.meta;
+package com.microservices.choreographyorchestrator.dto.definition;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class FailExecutionInfoDto {
+public class SuccessExecutionInfoDto {
     private String channelName;
 
     private String eventType;
