+++ src/main/java/com/microservices/choreographyorchestrator/repository/SagaStepTransitionRepository.java
@@ -1,7 +1,7 @@
 package com.microservices.choreographyorchestrator.repository;
 
-import com.microservices.choreographyorchestrator.domain.SagaStepDefinition;
 import com.microservices.choreographyorchestrator.domain.SagaStepTransitionEvent;
+import com.microservices.choreographyorchestrator.domain.definition.SagaStepDefinition;
 import org.springframework.data.neo4j.repository.Neo4jRepository;
 
 import java.util.List;
